# contracts/openapi/layer4-agents.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Layer 4: Agentic Workflow Engine",
    "description": "LangGraph-powered workflow orchestration for Value Fabric with multi-agent support",
    "version": "0.2.0"
  },
  "paths": {
...
    "/v1/accounts/{account_id}": {
      "get": {
        "tags": [
          "Accounts",
          "Accounts"
        ],
        "summary": "Get Account",
        "description": "Get full account detail.",
        "operationId": "get_account_v1_accounts__account_id__get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDetailSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/tenants/register": {
      "post": {
        "tags": [
          "layer4-agents"
        ],
        "summary": "POST /v1/tenants/register",
        "description": "Discovered from router source (value-fabric\\layer4-agents\\src\\tenants\\api\\routes\\registration.py).",
        "operationId": "discovered_post_v1_tenants_register",
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
...
    "/v1/tenant/settings": {
      "get": {
        "tags": [
          "layer4-agents"
        ],
        "summary": "GET /v1/tenant/settings",
        "description": "Discovered from router source (value-fabric\\layer4-agents\\src\\api\\routes\\frontend_compat.py).",
        "operationId": "discovered_get_v1_tenant_settings",
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      },
      "patch": {
        "tags": [
          "layer4-agents"
        ],
        "summary": "PATCH /v1/tenant/settings",
        "description": "Discovered from router source (value-fabric\\layer4-agents\\src\\api\\routes\\frontend_compat.py).",
        "operationId": "discovered_patch_v1_tenant_settings",
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/v1/auth/register": {
      "post": {
        "tags": [
          "layer4-agents"
        ],
        "summary": "POST /v1/auth/register",
        "description": "Discovered from router source (value-fabric\\layer4-agents\\src\\api\\routes\\frontend_compat.py).",
        "operationId": "discovered_post_v1_auth_register",
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountDetailSchema": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "provider": {
            "$ref": "#/components/schemas/CRMProvider"
          },
          "provider_record_id": {
            "type": "string",
            "title": "Provider Record Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "company_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Size"
          },
          "annual_revenue": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annual Revenue"
          },
          "headquarters": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headquarters"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "owner_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Name"
          },
          "owner_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Email"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "segment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segment"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "last_synced_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Synced At"
          },
          "sync_status": {
            "$ref": "#/components/schemas/SyncStatus"
          },
          "source_attribution": {
            "type": "string",
            "title": "Source Attribution",
            "description": "Human-readable sync info: 'Synced from Salesforce 2 hours ago'"
          },
          "provider_badge": {
            "type": "string",
            "title": "Provider Badge",
            "description": "Display badge: 'Salesforce' or 'HubSpot'"
          },
          "opportunities": {
            "items": {
              "$ref": "#/components/schemas/OpportunitySchema"
            },
            "type": "array",
            "title": "Opportunities",
            "default": []
          },
          "contacts": {
            "items": {
              "$ref": "#/components/schemas/ContactSchema"
            },
            "type": "array",
            "title": "Contacts",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "provider",
          "provider_record_id",
          "name",
          "created_at",
          "updated_at",
          "sync_status",
          "source_attribution",
          "provider_badge"
        ],
        "title": "AccountDetailSchema",
        "description": "Full account detail response."
      },
...
      "CRMProvider": {
        "type": "string",
        "enum": [
          "salesforce",
          "hubspot"
        ],
        "title": "CRMProvider",
        "description": "Supported CRM providers."
      },
...
      "ContactSchema": {
        "properties": {
          "provider_contact_id": {
            "type": "string",
            "title": "Provider Contact Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "is_primary": {
            "type": "boolean",
            "title": "Is Primary",
            "default": false
          },
          "last_synced_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Synced At"
          }
        },
        "type": "object",
        "required": [
          "provider_contact_id",
          "name",
          "last_synced_at"
        ],
        "title": "ContactSchema",
        "description": "Contact data embedded in account detail response."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OpportunitySchema": {
        "properties": {
          "provider_opportunity_id": {
            "type": "string",
            "title": "Provider Opportunity Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "value": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "probability": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Probability"
          },
          "close_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close Date"
          },
          "last_synced_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Synced At"
          }
        },
        "type": "object",
        "required": [
          "provider_opportunity_id",
          "name",
          "stage",
          "last_synced_at"
        ],
        "title": "OpportunitySchema",
        "description": "Opportunity data embedded in account detail response."
      },
...
      "SyncStatus": {
        "type": "string",
        "enum": [
          "synced",
          "pending",
          "failed",
          "stale"
        ],
        "title": "SyncStatus",
        "description": "Account sync status from CRM provider."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}